market_predictor/requirements.txt
# Core
requests>=2.31
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
joblib>=1.3
scipy>=1.12
python-dotenv>=1.0

# NLP / ML
sentence-transformers>=2.7
transformers>=4.40,<5.0
torch>=2.0
pdfplumber>=0.11
pypdf>=4.0

# Vector store
chromadb>=0.5

# Dashboard
streamlit>=1.35
plotly>=5.20

# API
fastapi>=0.111
uvicorn[standard]>=0.29

# AI commentary
google-generativeai>=0.5
anthropic>=0.25

# Utilities
textblob>=0.17
apscheduler>=3.10
